smart_attendance/requirements.txt
streamlit==1.36.0
opencv-python==4.10.0.84
face_recognition==1.3.0
dlib-bin
numpy==2.2.6
pandas==2.2.3
python-dotenv==1.0.1
supabase==2.4.6
Pillow==10.4.0